+++ guillotina/cookiecutter/application/{{cookiecutter.package_name}}/{{cookiecutter.package_name}}/tests/conftest.py
@@ -1,0 +1,4 @@
+pytest_plugins = [
+    'guillotina.tests.fixtures',
+    '{{cookiecutter.package_name}}.tests.fixtures'
+]