+++ guillotina/cookiecutter/application/{{cookiecutter.package_name}}/{{cookiecutter.package_name}}/tests/test_install.py
@@ -1,0 +1,7 @@
+import asyncio
+
+
+async def test_install({{cookiecutter.package_name}}_requester):  # noqa
+    async with {{cookiecutter.package_name}}_requester as requester:
+        response, _ = await requester('GET', '/db/guillotina/@addons')
+        assert '{{cookiecutter.package_name}}' in response['installed']